add OS signal handling code to VMachine
We need to select main VM for signal delivery and tear down signal
delivery when this VM goes out. This patch adds the required code.

The method added to VMachine leads to a no-op on Windows platform for
now. When we get some code that will actually be able to broadcast a
Windows equivalent of signals, it will go there.

Signed-off-by: Jan Dvorak <mordae@anilinux.org>

git-svn-id: svn://falconpl.org/falcon/core/trunk@1533 78e5f83a-d9f7-4e86-b854-dad15b5eb37a