CMake: Forward FALCON_ variables to the Falcon_ ones.
Some time ago FindBoost.cmake, one of the most advanced cmake modules,
introduced a new variable namig scheme that just makes sense.  When
  find_package(Foo)
is used in a cmake script, the corresponding variables to be used are
  Foo_LIBRARIES, Foo_INCLUDE_DIRS,..
This just makes more sense than the former convention to make it all
uppercase, no matter of it was
  find_package(Foo),
  find_package(FOO) or

Command line options are however still all uppercase
  -DFOO_*

The origin of this mess is cmake only allowing UPPERCASE for commands
and variables in older released.

This can however be used to our advantage, as users are trained to
pass UPPER_CASE options to the command line and our interal
Normal_CASE aren't overwritten by accident.