Cleaned up QVMContext, added isSleeping to check if queue is empty, added runningItem to check what item it is running by name.
start() requires a name (easier than figuring out a function from a item). start() accepts psteps for onComplete and onError of the item.

params are now used

NOTE: MUST call terminate() before deleting the VM or before decref the context. If not it will not stop running and the application will not close.